requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1

langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.5

chromadb==0.4.24
pypdf==4.2.0

pillow==10.3.0